adds endpoints for portfolio work; sets up deletion of portfolio work from filesystem on removal from db

+++ routes/servicePricingProfilesRouter/servicePricingRoutes/mixingPricingRouter.js
@@ -1,11 +1,14 @@
 const express = require('express');
 const MixingPricingProfile = require('../../../models/pricingProfiles/mixingPricingProfile');
+const cors = require('../../cors');
+const auth = require('../../../authenticate');
 
 const mixingPricingRouter = express.Router();
 
 mixingPricingRouter
 	.route('/')
-	.get((req, res, next) => {
+	.options(cors.corsWithOptions, (req, res) => res.sendStatus(200))
+	.get(cors.cors, (req, res, next) => {
 		MixingPricingProfile.find()
 			.then((profiles) => {
 				res.statusCode = 200;
@@ -14,8 +17,10 @@ mixingPricingRouter
 			})
 			.catch((err) => next(err));
 	})
-	.post((req, res, next) => {
-		MixingPricingProfile.create(req.body)
+	.post(cors.corsWithOptions, auth.verifyUser, (req, res, next) => {
+		const newProfile = req.body;
+		newProfile.userId = req.user._id;
+		MixingPricingProfile.create(newProfile)
 			.then((profile) => {
 				res.statusCode = 200;
 				res.setHeader('Content-Type', 'application/json');
@@ -23,8 +28,8 @@ mixingPricingRouter
 			})
 			.catch((err) => next(err));
 	})
-	.delete((req, res, next) => {
-		MixingPricingProfile.deleteMany()
+	.delete(cors.corsWithOptions, auth.verifyUser, (req, res, next) => {
+		MixingPricingProfile.deleteMany({ userId: req.user._id })
 			.then((response) => {
 				res.statusCode = 200;
 				res.setHeader('Content-Type', 'application/json');
@@ -93,7 +98,11 @@ mixingPricingRouter
 			.catch((err) => next(err));
 	})
 	.put((req, res, next) => {
-		MixingPricingProfile.findByIdAndUpdate(req.params.profileId, { $set: req.body }, { new: true })
+		MixingPricingProfile.findByIdAndUpdate(
+			req.params.profileId,
+			{ $set: req.body },
+			{ new: true }
+		)
 			.then((profile) => {
 				res.statusCode = 200;
 				res.setHeader('Content-Type', 'application/json');
